+++ apis/cloudflare/package.json
@@ -16,11 +16,6 @@
   "dependencies": {
     "@braintrust/proxy": "workspace:*",
     "@microlabs/otel-cf-workers": "1.0.0-rc.15",
-    "@opentelemetry/api": "^1.7.0",
-    "@opentelemetry/core": "^1.18.1",
-    "@opentelemetry/resources": "^1.18.1",
-    "@opentelemetry/sdk-metrics": "^1.18.1",
-    "@opentelemetry/semantic-conventions": "^1.18.1",
     "dotenv": "^16.3.1"
   }
 }
