+++ packages/proxy/src/metrics.ts
@@ -1,26 +1,36 @@
-import { MeterProvider } from "@opentelemetry/sdk-metrics";
-import { PrometheusRemoteWriteExporter } from "./exporter";
+import { MeterProvider, MetricReader } from "@opentelemetry/sdk-metrics";
 import { Resource } from "@opentelemetry/resources";
 import { v4 as uuidv4 } from "uuid";
 import opentelemetry from "@opentelemetry/api";
 
 let metricsInitialized = false;
 declare global {
-  var metricReader: PrometheusRemoteWriteExporter | undefined;
+  var metricReader: MetricReader | undefined;
 }
-export function safeInitMetrics(remoteWriteUrl: string) {
+export function safeInitMetrics(
+  metricReader: MetricReader,
+  resourceLabels?: Record<string, string>
+) {
   if (metricsInitialized) {
     return;
   }
 
-  globalThis.metricReader = new PrometheusRemoteWriteExporter({
-    url: remoteWriteUrl,
-  });
+  globalThis.metricReader = metricReader;
 
+  // DEVNOTE: This means that each request will be its own instance, which will explode
+  // timeseries cardinality in Prometheus. This is probably okay, unless we have a really significant
+  // number of requests (10k+ per second, assuming we have around 20 metrics, according to
+  // https://prometheus.io/docs/prometheus/1.8/storage/#settings-for-high-numbers-of-time-series)
+  //
+  // To solve this, we'll have to partially aggregate metrics,
+  // e.g. in a durable worker or an aggregating push gateway (e.g. https://github.com/zapier/prom-aggregation-gateway).
+  // Because the remote_write format collapses everything down to simple metrics, we can likely do this
+  // in terms of the basic timeseries format.
   const resource = Resource.default().merge(
     new Resource({
       job: "braintrust-proxy",
       instance: uuidv4(),
+      ...resourceLabels,
     })
   );
 
@@ -81,4 +91,8 @@ export function exponentialBuckets(
     start *= factor;
   }
   return buckets;
+}
+
+export function nowMs() {
+  return performance?.now ? performance.now() : Date.now();
 }
