+++ packages/proxy/src/proxy.ts
@@ -28,6 +28,7 @@ import {
   anthropicCompletionToOpenAICompletion,
   anthropicEventToOpenAIEvent,
 } from "./providers/anthropic";
+import { getMeter, nowMs } from "./metrics";
 
 interface CachedData {
   headers: Record<string, string>;
@@ -59,6 +60,12 @@ export async function proxyV1(
   cachePut: (encryptionKey: string, key: string, value: string) => void,
   digest: (message: string) => Promise<string>
 ): Promise<void> {
+  const meter = getMeter("proxy-metrics");
+
+  const cacheHits = meter.createCounter("results_cache_hits");
+  const cacheMisses = meter.createCounter("results_cache_misses");
+  const cacheSkips = meter.createCounter("results_cache_skips");
+
   proxyHeaders = Object.fromEntries(
     Object.entries(proxyHeaders).map(([k, v]) => [k.toLowerCase(), v])
   );
@@ -156,7 +163,9 @@ export async function proxyV1(
   let stream: ReadableStream<Uint8Array> | null = null;
   if (useCache) {
     const cached = await cacheGet(encryptionKey, cacheKey);
+
     if (cached !== null) {
+      cacheHits.add(1);
       const cachedData: CachedData = JSON.parse(cached);
 
       for (const [name, value] of Object.entries(cachedData.headers)) {
@@ -170,7 +179,11 @@ export async function proxyV1(
           controller.close();
         },
       });
+    } else {
+      cacheMisses.add(1);
     }
+  } else {
+    cacheSkips.add(1);
   }
 
   if (stream === null) {
